:sparkles: MongoDBのハンドラを変更
MongoDBへの接続を管理するハンドラを外部から注入できるように変更した。
Refer to #3
