+++ api/app/internal/infrastructure/mongodb/db.go
@@ -1,0 +1,29 @@
+package mongodb
+
+import (
+	"context"
+	"log"
+	"time"
+
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+	"go.mongodb.org/mongo-driver/mongo/readpref"
+)
+
+// MongoDBのハンドラを生成する。
+//
+// クライアントの初期化に失敗した場合は異常終了する。
+// MongoDB への疎通確認に失敗した場合は異常終了する。
+func NewMongoDatabase(uri, name string) *mongo.Database {
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	client, err := mongo.Connect(ctx, options.Client().ApplyURI(uri))
+	if err != nil {
+		log.Fatalf("Failed to initialize the client of MongoDB: %v", err)
+	}
+	if err := client.Ping(ctx, readpref.Primary()); err != nil {
+		log.Fatalf("Failed to ping MongoDB: %v", err)
+	}
+	db := client.Database(name)
+	return db
+}
